+++ src/main/java/com/teamnine/humanofdelivery/dto/review/ReviewRequestDto.java
@@ -1,4 +1,4 @@
-package com.teamnine.humanofdelivery.dto;
+package com.teamnine.humanofdelivery.dto.review;
